Use service=s3 by default for DO Spaces and Linode
A few services expect the `service` parameter to be set to `s3` where
this is not currently inferred, such that "default" aws4fetch use fails
with authentication errors. We already handle this case for R2 and B2,
but we were missing Linode Object Storage and DigitalOcean Spaces:

https://github.com/mhart/aws4fetch/issues/15
https://github.com/twilligon/git-lfs-s3-proxy/issues/10

Handle `service` as we do for R2 and B2 so aws4fetch works by default.
Also, add virtual-host and path-style URL tests for both providers.